Add ReplacementOptions to filter telemetry events
For node application insights clients, there is a new constructor
parameter that allows users to pass a list of replacement options.

Each option includes a lookup regular expression and a replacement
string.

For each key in the base data or the envelope that matches the
lookup, the value will be replaced with the associated string. If
the value is undefined, then the key will be deleted.

Note that this only adds replacements for node clients as I did not
see an easy way to implement the same in web clients.