Improve properties capture for an error event
`sender.sendErrorData` doesn't always receive data object with a properties
field. In the case of unhandled errors vscode itself calls this with all the
properties being in the top level of the `data` argument. This PR accounts for
that and fallbacks to using full data object if a properties field doesn't exist