Set xhr override before creating AppInsights client
If not supplied to the constructor the delegate AppInsightsCore class
does not pick up the default channel (`BreezeChannelIdentifier =
"AppInsightsChannelPlugin"`) plugin configuration and therefore doesn't
use the supplied XHR override.

This is subtle to notice when running in Node 18+ (VS Code 1.82+) since
the AppInsights SDK mostly uses the `fetch` API which is available in
Node and appears to work.